move filtering functions out of scatterplot
